Cleared persistent wallpaper repo value after reading it
This CL aims at clearing persistent wallpaper repo value after we have
used it. This ensures that we fix the bug where we kept on seeing emoji
wallpaper everytime wallpaper preview is opened with any wallpaper
especially in the case when emoji wallpaper is opened for the first
time. Also persisting wallpaper on app configuration restarts used to
rely on persistent wallpaper repo too, and it no longer worked when the
value is cleared on read. Instead of relying on persistent wallpaper
repo to get the previous wallpaper model, get the value last stored in
the preview view model.

Bug: 381780002
Flag: EXEMPT bug
Test: Tested by checking if we see the right preview each time
Test: Tested by rotating in preview screens to check wallpaper persists
Change-Id: Ib824d541475f9954fdfdf31e3994e2b44defa3bf