Make sure the right lifecycle scope for the binder
When transition in between fragment, the fragment remains while the view
is recreated. Thus we should always use the view's lifecycle scope for
listening to the data flows.

Not doing so will make the fragment listen to as many data flows as the
times we go back and forth the 2 fragments.

Test: Manully tested the issue was gone
Fixes: 383824030
Flag: com.android.systemui.shared.new_customization_picker_ui
Change-Id: I2b26fe18eb7ca2fdb5844df6b10475dbf78fb6f4