Selectively use separate preview engines for lock and home previews
When previewing a wallpaper using the side-by-side previews in
SmallPreviewBinder, a single engine is used for both lock and home
screens. This causes issues when a single wallpaper shows different
content for each screen.

This changes selectively uses different preview engines for each, and
propagates the preview screen (which is different than the screen the
wallpaper is currently applied on) through to the wallpaper connection.

Fixes: 389962381
Bug: 390731022
Flag: EXEMPT bugfix
Test: manually verified correct behavior, see bug
Change-Id: I61a25f16e35472b4aa2e5f432f616fd271fdf879