Update previews on wallpaper change
Create a new WallpaperRepository2 for the refactored preview to
introduce the BroadcastDispatcher in the constructor. This is difficult
with the WallpaperRepository since it is used in older picker
infrastructure and created manually rather than injected in many tests.
Subscribe to wallpaper change in WallpaperRepository2 by creating a
broadcast flow that listens to the wallpaper changed intent, then get
updated wallpapers when the flow emits. Also added test and fixed
BroadcastDispatcherTest.

Flag: com.android.systemui.shared.new_customization_picker_ui
Test: manually verified by changing to static and live wallpapers
Test: added unit tests
Bug: 288312530
Change-Id: I8186b6d16453b7cae28c3af80ce34a90b1156efd