Fix frozen selection animation in grid picker
Partial fix to the grid picker bug where the selected option is not
updated after a selection is made, and the animation is frozen. This
fixes the frozen animation, by properly cancelling the animation when a
new value is received and a new animation is started.

Flag: EXEMPT bug fix
Bug: 381026027
Bug: 386309338
Test: manually verified by changing options in grid picker
Change-Id: Ief19815ef262a436904b234230819511ff3b13d0