Improve pager preview extra space
To avoid extra space in between the lock and home previews, we need to
make labels and the AspectRatioFrameLayout as siblings working in the
same Motion Layout.

The only downside is that we lose the show and hide labels when entering
the secondary screen. Since this is a more minor improvements that 2
previews are too far apart, we prioritize this change and will seek for
a soltuion to properly hide the labels when entering secondary screen.

Test: Manully tested.
Bug: 398270437
Flag: com.android.systemui.shared.new_customization_picker_ui
Change-Id: Ifc9beef9e65aa7b89d3d63b44cdc2aebe42a397f