Added changes for photopicker
This CL aims at adding changes for photo picker integration. This means
that users who are trying to select a custom photo to be set as
wallpaper can now use the Android Photopicker to select those photos.
This is behind a flag so we can turn it off as needed.

Bug: 399381227
Flag: com.android.systemui.shared.new_customization_picker_ui
Test: Tested by building and installing picker and checking if clicking
on launch photos button opens photo picker. See bug for screen-record.

Change-Id: Ied336df1aaf5daa7bfa711e02cc85b3d5af32ba0