Fix color option list crash
Crash was occuring occassionally in the color option recycler view
adapter, because the color options in the view model can emit multiple
times on app start, but the binder starts a new coroutine each time to
run diffUtils on the old and new list and dispatch results to the
adatper, without cancelling the old coroutine. Therefore if color
options emits twice in a row, it could lead to the adapter being
notified to insert the options twice, and result in an inconsistent
state.

Flag: com.android.systemui.shared.new_customization_picker_ui
Bug: 394380893
Test: manually verified by repeatedly entering picker app
Change-Id: I63a0bb340f11988f6b589923ef6fec275770c77e