Min preview height
We will do the best to fit the whole option container on the screen with
the preview visible. However, there are cases that the preview can be
pushed up and gets to small.
We specify minimum expanded and collapsed height in case the screen can
no host the whole option container.

Test: Manaully tested. See bug.
Bug: 392019188
Bug: 388411232
Flag: com.android.systemui.shared.new_customization_picker_ui
Change-Id: I0d1be9b9265515b1879202845952b8fff0dd264f