Changes for launching activity using pending intent
This CL aims at ensuring that clicking on the sign in button can allow
user to launch the sign in flow that is done by using pending intent to
launch a new activity.

Bug: 403670498
Flag: com.android.systemui.shared.new_customization_picker_ui
Test: Tested by building and installing picker on local, clicking on
sign in button launches the sign in activity.

Change-Id: I6c338cd8051eff3ce385f8135247fe5e8d88e9b8